+++ js/main.js
@@ -1,4 +1,4 @@
-import {createPhotoDescriptions} from './create-photo-descriptions.js';
+import {createPhotoDescriptions} from './data.js';
 
 // eslint-disable-next-line no-console
 console.log(createPhotoDescriptions());
